Conflicts:
	Zend/zend.h
	Zend/zend_API.c
	Zend/zend_compile.c
	Zend/zend_language_parser.y
	Zend/zend_language_scanner.c
	Zend/zend_multiply.h
	Zend/zend_operators.c
	Zend/zend_operators.h
	Zend/zend_types.h
	Zend/zend_vm_def.h
	Zend/zend_vm_execute.h
	configure.in
	tests/lang/operators/bitwiseShiftLeft_basiclong_64bit.phpt
	tests/lang/operators/bitwiseShiftLeft_variationStr.phpt
	tests/lang/operators/bitwiseShiftLeft_variationStr_64bit.phpt
	tests/lang/operators/bitwiseShiftRight_basiclong_64bit.phpt
	tests/lang/operators/bitwiseShiftRight_variationStr.phpt
	win32/build/config.w32

+++ Zend/zend_exceptions.c
@@ -363,7 +363,7 @@ ZEND_METHOD(error_exception, getSeverity)
 
 
 #define TRACE_ARG_APPEND(vallen) do { \
-		int len = str->len; \
+		size_t len = str->len; \
 		str = zend_string_realloc(str, len + vallen, 0); \
 		memmove(str->val + len - l_added + 1 + vallen, str->val + len - l_added + 1, l_added); \
 	} while (0)
@@ -628,10 +628,10 @@ ZEND_METHOD(exception, getPrevious)
 	RETURN_ZVAL(previous, 1, 0);
 } /* }}} */
 
-int zend_spprintf(char **message, int max_len, const char *format, ...) /* {{{ */
+size_t zend_spprintf(char **message, size_t max_len, const char *format, ...) /* {{{ */
 {
 	va_list arg;
-	int len;
+	size_t len;
 
 	va_start(arg, format);
 	len = zend_vspprintf(message, max_len, format, arg);
@@ -640,7 +640,7 @@ int zend_spprintf(char **message, int max_len, const char *format, ...) /* {{{ *
 }
 /* }}} */
 
-zend_string *zend_strpprintf(int max_len, const char *format, ...) /* {{{ */
+zend_string *zend_strpprintf(size_t max_len, const char *format, ...) /* {{{ */
 {
 	va_list arg;
 	zend_string *str;
